added status codes

+++ CinemaApi/Controllers/MoviesController.cs
@@ -1,5 +1,6 @@
 ﻿using CinemaApi.Data;
 using CinemaApi.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using System.Collections.Generic;
 
@@ -23,45 +24,67 @@ namespace CinemaApi.Controllers
 
     // GET: api/<MoviesController>
     [HttpGet]
-    public IEnumerable<Movie> Get()
+    public IActionResult Get()
     {
-      return _dbContext.Movies;
+      return Ok(_dbContext.Movies);
+      //return _dbContext.Movies;
+      //return StatusCode(StatusCodes.Status200OK);
     }
 
     // GET api/<MoviesController>/5
     [HttpGet("{id}")]
-    public Movie Get(int id)
+    public IActionResult Get(int id)
     {
       var movie = _dbContext.Movies.Find(id);
-      return movie;
+      if (movie == null)
+      {
+        return NotFound("No record found against this Id");
+      }
+      return Ok(movie);
     }
 
     // POST api/<MoviesController>
     [HttpPost]
-    public void Post([FromBody] Movie movieObj)
+    public IActionResult Post([FromBody] Movie movieObj)
     {
       _dbContext.Movies.Add(movieObj);
       _dbContext.SaveChanges();
+      return StatusCode(StatusCodes.Status201Created);
     }
 
     // PUT api/<MoviesController>/5
     [HttpPut("{id}")]
-    public void Put(int id, [FromBody] Movie movieObj)
+    public IActionResult Put(int id, [FromBody] Movie movieObj)
     {
       var movie = _dbContext.Movies.Find(id);
-      movie.Name = movieObj.Name;
-      movie.Language = movieObj.Language;
-      _dbContext.SaveChanges();
-
+      if (movie == null)
+      {
+        return NotFound("No record found against this Id");
+      }
+      else
+      {
+        movie.Name = movieObj.Name;
+        movie.Language = movieObj.Language;
+        _dbContext.SaveChanges();
+        return Ok("Record updated successfully");
+      }
     }
 
     // DELETE api/<MoviesController>/5
     [HttpDelete("{id}")]
-    public void Delete(int id)
+    public IActionResult Delete(int id)
     {
       var movie = _dbContext.Movies.Find(id);
-      _dbContext.Movies.Remove(movie);
-      _dbContext.SaveChanges();
+      if (movie == null)
+      {
+        return NotFound("No record found against this Id");
+      }
+      else
+      {
+        _dbContext.Movies.Remove(movie);
+        _dbContext.SaveChanges();
+        return Ok("Record deleted");
+      }
     }
   }
 }
